+++ hephaistos/backups.py
@@ -30,10 +30,10 @@ def __get_file(file: Path) -> Path:
     return BACKUP_DIR.joinpath(file)
 
 
-def invalidate() -> None:
+def discard() -> None:
     if BACKUP_DIR.exists():
         dir_util.remove_tree(str(BACKUP_DIR))
-    LOGGER.info(f"Invalidated backups at '{BACKUP_DIR}'")
+    LOGGER.info(f"Discarded backups at '{BACKUP_DIR}'")
 
 
 def restore() -> None: