+++ hephaistos/hashes.py
@@ -34,10 +34,10 @@ def __get_file(file: Path) -> Path:
     return HASH_DIR.joinpath(file).with_suffix(HASH_FILE_EXTENSION)
 
 
-def invalidate() -> None:
+def discard() -> None:
     if HASH_DIR.exists():
         dir_util.remove_tree(str(HASH_DIR))
-        LOGGER.info(f"Invalidated hashes at '{HASH_DIR}'")
+        LOGGER.info(f"Discarded hashes at '{HASH_DIR}'")
 
 
 class HashMismatch(LookupError): ...