+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   Support `--force` in interactive mode to allow users to repatch after a game update.
+
 ## [v0.4.0] - 2021-08-15
 
 ### Added